requirements.txt
openai
langchain
chromadb
sentence-transformers
pypdf2
streamlit
python-dotenv
langchain-community
pypdf